exception message added

+++ EmployeeCurdOperation/src/main/java/com/emp/service/EmployeeServiceImpl.java
@@ -8,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.emp.entity.Employee;
+import com.emp.exception.NoEmpFoundException;
 import com.emp.repository.EmployeeRepository;
 
 @Service
@@ -18,12 +19,13 @@ public class EmployeeServiceImpl implements EmployeeService{
 	
 	@Override
 	public List<Employee> getAllEmployees() {
+		
 		return employeeRepository.findAll();
 	}
 
 	@Override
 	public Employee getEmployeebyId(long id) {
-		return employeeRepository.findById(id).orElse(null);
+		return employeeRepository.findById(id).orElseThrow(()->new NoEmpFoundException("No Emp found with id : "+id));
 	}
 
 	@Override
@@ -33,13 +35,24 @@ public class EmployeeServiceImpl implements EmployeeService{
 
 	@Override
 	public Employee updateEmployee(Employee employee) {
+		
+		long empId=employee.getId();
+		
+		if(!employeeRepository.existsById(empId)) {
+			throw new NoEmpFoundException("No emp found with id : "+empId+" record cannot update");
+		}
+		
 		return employeeRepository.save(employee);
 	}
 
 	@Override
 	@Transactional
 	public void deleteEmployee(long id) {
-        employeeRepository.deleteById(id);
+        //check emp exist or not 
+		if(!employeeRepository.existsById(id)) {
+			throw new NoEmpFoundException("No Emp found for delete with id : "+id);
+		}
+		employeeRepository.deleteById(id);
 		
 	}
